Added simplex plugin
