+++ src/modules/UI/components/ControlPanel/ControlPanel.ui.js
@@ -3,17 +3,23 @@
 import React, { Component } from 'react'
 import { Image, TouchableHighlight, View } from 'react-native'
 import MDIcon from 'react-native-vector-icons/MaterialIcons'
+import { Actions } from 'react-native-router-flux'
+import { sprintf } from 'sprintf-js'
 
 import person from '../../../../assets/images/sidenav/accounts.png'
+import buysell from '../../../../assets/images/sidenav/buysell.png'
 import { emptyGuiDenomination } from '../../../../types'
 import type { GuiDenomination } from '../../../../types'
 import { getDenomFromIsoCode } from '../../../utils.js'
 import T from '../../components/FormattedText'
+import s from '../../../../locales/strings.js'
 import ExchangeRate from '../ExchangeRate/index.js'
 import Gradient from '../Gradient/Gradient.ui'
 import SafeAreaView from '../SafeAreaView/SafeAreaViewDrawer.ui.js'
 import Main from './Component/MainConnector'
 import styles from './style'
+
+const PLUGIN_BUYSELL_TEXT = sprintf(s.strings.title_plugin_buysell)
 
 export type Props = {
   currencyLogo: string,
@@ -77,7 +83,7 @@ export default class ControlPanel extends Component<Props> {
             />
           </View>
           <TouchableHighlight onPress={this._handlePressUserList} underlayColor={styles.underlay.color}>
-            <View style={styles.user.container}>
+            <View style={[styles.user.container, styles.others.borderBottom]}>
               <View style={styles.iconImageContainer}>
                 <Image style={styles.iconImage} source={person} />
               </View>
@@ -85,6 +91,22 @@ export default class ControlPanel extends Component<Props> {
               <MDIcon style={styles.icon} name={arrowIcon} />
             </View>
           </TouchableHighlight>
+
+          {!this.props.usersView && (
+            <TouchableHighlight
+              style={styles.others.iosTouchableHighlight}
+              underlayColor={styles.main.iosTouchableHighlightUnderlayColor}
+              onPress={Actions.buysell}
+            >
+              <View style={[styles.others.link, styles.others.borderBottom, { flex: 1 }]}>
+                <View style={styles.iconImageContainer}>
+                  <Image style={styles.iconImage} source={buysell} />
+                </View>
+                <View style={styles.others.textContainer}>
+                  <T style={styles.others.text}>{PLUGIN_BUYSELL_TEXT}</T>
+                </View>
+              </View>
+            </TouchableHighlight>)}
           <Main />
         </Gradient>
       </SafeAreaView>
